moved time to its own module
